NOTE - CLEAN BUILD REQUIRED! ASH ints are now Java longs. ASH floats are now Java doubles. ASH booleans are now Java BigBoo... no wait, they were fine as they are. There is no increase in runtime memory requirements; since no Value needs both int and float content, they can be both stored in the same field. Known limitations at this point: * There still isn't a way to get substat or meat amounts beyond 2 billion, but at least that's now a possibility. * Ints outside the 32-bit range, and floats outside the single-precision range, probably cannot be stored in a map file at the moment, nor survive round-trip conversion to strings. * Ints outside the 32-bit range, when passed to most library functions, will simply have the top 32 bits ignored, rather than doing any range checking on the full 64-bit value (so 2**32+7 is exactly the same as 7 in most cases).
git-svn-id: http://svn.code.sf.net/p/kolmafia/code@11059 b29ace70-8910-0410-8dcc-aa2fc6433167

+++ src/net/sourceforge/kolmafia/textui/DataTypes.java
@@ -166,7 +166,7 @@ public class DataTypes
 	public static final Value FALSE_VALUE = new Value( false );
 	public static final Value ZERO_VALUE = new Value( 0 );
 	public static final Value ONE_VALUE = new Value( 1 );
-	public static final Value ZERO_FLOAT_VALUE = new Value( 0.0f );
+	public static final Value ZERO_FLOAT_VALUE = new Value( 0.0 );
 	public static final Value MUSCLE_VALUE = DataTypes.STAT_VALUES[0];
 	public static final Value MYSTICALITY_VALUE = DataTypes.STAT_VALUES[1];
 	public static final Value MOXIE_VALUE = DataTypes.STAT_VALUES[2];
@@ -390,7 +390,7 @@ public class DataTypes
 		{
 			if ( name.equalsIgnoreCase( DataTypes.STATS[ i ] ) )
 			{
-				return STAT_VALUES[ i ].intValue();
+				return (int) STAT_VALUES[ i ].intValue();
 			}
 		}
 		return -1;
@@ -619,16 +619,16 @@ public class DataTypes
 		return val ? ONE_VALUE : ZERO_VALUE;
 	}
 
-	public static final Value makeIntValue( final int val )
+	public static final Value makeIntValue( final long val )
 	{
 		return val == 0 ? ZERO_VALUE :
 		       val == 1 ? ONE_VALUE :
 		       new Value( val );
 	}
 
-	public static final Value makeFloatValue( final float val )
+	public static final Value makeFloatValue( final double val )
 	{
-		return val == 0.0f ? ZERO_FLOAT_VALUE : new Value( val );
+		return val == 0.0 ? ZERO_FLOAT_VALUE : new Value( val );
 	}
 
 	public static final Value makeBooleanValue( final int num )
